Add files via upload
APMC management system stores the details about the 
farmers, dealers, auctions performed with the dealers and the 
final dealing records . Farmer is the main charge as he sells the 
products in large quantity (in orders of quintals ) from him
the complete process starts. 
 The project is completely programmed using Code blocks 
where the inheritance concept of C++ (including Multiple 
inheritance) for the different classes .
 File handling concept is used for storing the complete process 
details (in systematic manner) where these will be fed into file 
and stored which will be used for future use as more records 
can be appended.